+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## _Under development_
 
+- Modules
+    - Audit
+        - Path traversal (`path_traversal`)
+            - Updated `/etc/passwd` signatures to be more generic.
+
 ## 0.4.4 _(August 10, 2013)_
 
 - Options